optimizations
git-svn-id: https://svn.code.sf.net/p/speed-dreams/code/trunk@2540 30fe4595-0a0c-4342-8851-515496e4dcbd

Former-commit-id: 22deb98729cfc8d6786c5ae3cd7ebb2fd23b88dd
Former-commit-id: 80ac2937411fedad8f2fd49ad29dfe0b5742a2e0

+++ src/modules/simu/simuv3/differential.cpp
@@ -233,7 +233,7 @@ SimDifferentialUpdate(tCar *car, tDifferential *differential, int first)
                 float bias = differential->dTqMax * 0.5f* t_spin;
                 if (fabs(2.0*bias) < differential->dTqMin) 
                 {
-	            	bias = SIGN(bias) * 0.5f * differential->dTqMin;
+	            	bias = (float)(SIGN(bias) * 0.5f * differential->dTqMin);
 	            } 	  	
                 float open = 1.0f - fabs(pressure);
                 //DrTq0 = DrTq*(0.5f + bias) + spiderTq;
